requirements.txt
attrdict==2.0.1
base58==2.1.0
certifi==2020.12.5
chardet==4.0.0
cytoolz==0.11.0
ecdsa==0.16.1
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
greenlet==1.1.0
hexbytes==0.2.1
idna==2.10
importlib-metadata==4.0.1
Naked==0.1.31
parsimonious==0.8.1
plyvel==1.3.0
protobuf==3.15.8
# psycopg2==2.8.6
psycopg2-binary==2.8.6
pycryptodome==3.10.1
pysha3==1.0.2
PyYAML==5.4.1
requests==2.25.1
rlp==1.2.0
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.4.15
toolz==0.11.1
tronapi==3.1.6
trx-utils==1.0.4
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1